Moving more worker state management into its own class
